Dynamic Acceleration and Clearable Hills
Added in code that allows the train to accelerate and decelerate slower then faster (close to quadratic) based on how long it's been at its current speed. This code doesn't apply when the train is unfueled and going down hills, so it doesn't screech to a stop. Values can be changed to tweak how it feels, but right now (hopefully) nothing should break by doing that. 

There is an issue of Time.deltaTime being factored into the code twice every frame, but removing it makes everything so fast it doesn't make any sense, so it's staying, sorry mom.

There is now code that allows the train to never be pushed back down a hill if it's actively trying to climb it (aka, it's fueled).