Continuing WIP on track swapping
New approach - we're accessing the variables inside VertexPath from PathCreator using the GetVertexPath method. Now the code that switches what tracks the trains are on is located in trackLink.cs. The issue right now is a sticky one, though.

Basically, we need to know when a cart has reached the end of a path. We were using the variable "t" within one of VertexPath's methods to do this (because it was how all the other EndOfPathInstruction variables was handled), but it doesn't really seem to do what we want it to do sadly. 

So for now, we need to find a way to know when an individual cart has reached the end of a path for our approach to work. Sweet dreams.