WORKING Swap to NEXT TRACK
Ladies and gentlemen... it works. And it appears to work perfectly (for now). I have a whole lot of refactoring, documenting, and cleaning up to do. It works. I am not 100% positive why it works, but it works. Thank goodness. Thank goodness.

Here's the lowdown. Instead of mimicking the "loop" command for endOfLineInstruction, we're going to copy the "stop" instructions. This was found on accident, but instead of teleporting us back to the beginning, this instruction will keep us exactly where we are, and provide a smooth transition. My idea to use VertexPath.Length worked after all. I am emotional.

Left to do: 
- Clean up *all* the unnecessary crap I have thrown into this code in an attempt to make it work. All I need is the changes to PathFollower, the ability to pull Length from VertexPath at any given time, and my pathLink script (which I'm going to rename to PathLink for consistency's sake... wish me luck). 

Next stop? Adding in support for multiple carts (haha that's scary), adding in an actual changeable track (Unity is being stubborn about that rn), and finally handling moving in both directions (scary, but I have an idea).