WIP Multi-Train Workaround for Multi Tracks
I am so tired... but progress has been made. Right now there is an issue when using multiple trains. I have coded in a "forward offset" that always makes the very last train the one that starts at the beginning of a track.

There are two issues with this right now. DistanceTravelled doesn't like to be changed by this because it explodes every time I do it. 

Also, the so-called genius fix of using the "stop" behavior seems to backfire with multiple carts. I need to investigate further. Will report back. Now I sleep.