Multi-Carts so Close it's Illegal
I have an implementation of our new Track Swap that works for multi carts... with one hiccup. 

There is currently a bug where the distance between two carts is not kept consistent when swapping tracks. I believe I know why - it varies based on when the carts are actually swapping. The threshold for swapping is when a cart's distanceTravelled is greater than or equal to the distance of the path. The carts position is not updated by a constant value, so they all swap over to 0 on the next track at a different time, because they all cross that threshold at slightly different times! This needs a fix... in the next push!