The Dilemma
We grow closer and closer. The issue with inconsistencies in the code is now a problem of Time.deltaTime. The issue is that the rotationCalculator amount that is added to our speed in PathFollower is not equalized by Time.deltaTime... but we don't know how to equalize it yet. Because of this, the train reaches wildly different speeds depending on what resolution it runs at (aka, the framrate). We need a consistent Time.deltaTime.