We did it! Now, it's time to suffer!
Good news, pard! The train IS finally frame consistent! No matter the resolution, the train moves the same distance (with a very very tiny variance). 

Now for the new issue - track swapping still does not work. There is a reason for this - but it is consistent between going forwards and backwards, which is good. If we solve one, it'll solve the other.

Originally, the reason why there was backwards stutter was because of two things: one, there was a subtraction instead of an addition in the code. And two, we move slower backwards than we do forwards, and that exposes the issue.

When we move to swap tracks, we have to enter a threshold of distanceTravelled + 0.1 being greater than trackLength. The only problem with that, is at lower speeds where we update our position by smaller values, we can be in that range without having made it to the end. This causes our offset on the next track to be negative, holding us in place for a frame or two. That's how you get stutter.

Our idea to combat this was to take a proactive approach. On any frame, we check to see if distanceTravelled + amount is greater than trackLength. If so, on the NEXT frame we swap to the next track and set our offset value to distanceTravelled - prevTrack.trackLength. Then, we set our distanceTravelled to our offset value, therefore placing us the distance we have gone past the track.

This way we are ready to swap on the frame we need to, because we always check the frame before. It should work. It doesn't. Something weird is going on, and the math just straight up isn't working (you can watch it not work in the editor). Life is pain. I am suffering. I'm gonna sleep on it. Suffering 100 not chungus everybody disliked that reddit.com FREE stanger things season 4 download https://omfgdogs.com