I put a player in the game
The player is just a cylinder, but he's a good boy (for now). His name is Dylan.

The player is quite big compared to the train, but that can be fixed a number of ways (bigger train, smaller player, etc.). Also because programming is funny, if you jump as the player, you also stop the train hahahahahahahahahahahahahahahaha. Easy fix, don't care to do it now though.

I made both our old train camera and our new player object into prefabs that can be put into the scene and enabled/disabled however one pleases. Since both items are technically a camera, only one should be enabled at a time.

For the player, if you start the game, you will notice that the cursor disappears, meaning you will not be allowed to use the editor while the game is playing... is what I WOULD SAY if I was a stupid dumb idiot brain. Luckily, I added some code that allows the player to left click to get their mouse cursor back, if this is done, the user also loses control of their player object so no funky business happens while using the editor. You can also left click on the game window to regain control of the player. Very cool very epic definitely not thinking about how the train itself is still broken no siree