Small Tweaks and Branching Path
Added a couple of very small tweaks to the PathFollower code so that our train only removes itself from the trainsOnTrack array before it has passed (was causing an issue).

Also added in Tracks3, our *third* path, as a way to demonstrate track swapping (the whole freakin' reason we did what we did for the last two weeks). Now if you left click while on Tracks2 you will swap over to Tracks3 instead of Tracks when you get there.  You can't swap if you're already changing tracks, or if you're not on a track with a branching path. It works well, only issue right now is that the start point for Tracks3 is really weird, and updating its XYZ values to the position of other track end pieces is not consistent. Not really sure what's up with that. 

But anyways, it's nice to have a push that contains stuff that actually worked for once. That's pretty cool. 

I'm pushing this to master because of the very important tweak in PathFollower's oop when changing tracks. Further pushes will be in their own branch and focus on going in both directions on tracks.