Working Variable Direction (wow)
So Jackson had an extra gamer idea. Why don't we just put another track down on our tracks that goes in the other direction? Epic.

So we did that. It works. Some of the track pieces are stinky because placing these things is a total hassle, but it works! Wow! What next?