New Demo Area and Models
Created a large and in-charge new demo area for continued work which includes slopes, track swapping, and multi-directional paths. 

Also added in some new models as prefabs that we can populate that map with, like some buildings, hills, and trees. I've made a little town, started work on a forest, and added a big hill in.