New Train Controller and Updated UI
Did a little refactoring here. Now the train is controlled by a TrainController object, and so is every track (in the future when we implement enemy trains, they will need to override the commands of the TrainController that the player is using). This factored out a lot of the inputs from PathFollower and also allowed me to add in a new UI for controlling the train. You can go top speed, low speed, or no speed. There are brakes (which don't work, RIP idk why probably lerping). And you can change which track you're going to this way as well!

I also made the forest a little larger. Reminder: hills. Obscure areas with hills to make your world bigger. That is all. Party on party people.