Player Movement Baseline
We have a beginning for a player on the train! With a few tweaks to Jackson's original model, the player can now move and jump on the train with no problem. We slowed him down, too. He still needs a lot of fixing (general speed, jump height, crappy crouch, arrow key movement, etc), but he's on the way.