WIP Invisible Walls That Keep Player On Train
So this mostly works. Only problem is that the player doesn't glide off the walls if he touches them. Likewise, if the player jumps into a wall and continues to input, they'll just float there. Why does this happen? Idk. I googled this and found nothing so I'm going to push it because everything I've tried got me the same result, and that result is pain